商户列表完成

+++ shangou-pojo/src/main/java/com/lh/shangou/pojo/entity/ApprovalLog.java
@@ -34,7 +34,7 @@ public class ApprovalLog implements Serializable {
     /**
      * 审核人id
      */
-    private Integer approvalUserId;
+    private Long approvalUserId;
 
     /**
      * 审核人真实名字
